refactor(core): move canonicalDir where the cache decision can reach it
The shared-data decision has to compare `~/.nx` against the shared container
the same way the socket tier does, and `canonicalDir` lived in
`daemon/tmp-dir`, which imports `cache-directory` rather than the other way
round. It moves to `owned-private-dir`, alongside the guards that already use
this spelling.

Placed above the sticky-bit constant, so `isSafeSharedRoot`'s contract stays
directly above `isSafeSharedRoot` rather than ending up as the second of two
stacked doc blocks on a path helper.

`realpathSync` and `basename` were the only things `tmp-dir` still imported for
it, so both go with it.
